# backend/src/main/resources/api.yaml
openapi: 3.1.0
info:
  title: Fullstack Template Backend
  description: A small sample for a generic backend to showcase a fullstack application
  version: 1.0.0
servers:
  - url: 'http'
paths:
  /notes:
    get:
      tags:
        - Notes
      summary: Alle Notizen abrufen
      operationId: getNotes
      responses:
        '200':
          description: Liste aller Notizen
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Note'
    put:
      tags:
        - Notes
      summary: Adds a new note
      operationId: addNewNote
      requestBody:
        $ref: '#/components/requestBodies/NoteRequest'
      responses:
        '201':
          description: Note has been created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NoteCreated'
  /notes/{id}:
    delete:
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      tags:
        - Notes
      summary: deletes a specific note
      operationId: deleteNote
      responses:
        '204':
          description: Note deleted

components:
  schemas:
    Note:
      type: object
      required:
        - title
        - content
      properties:
        title:
          type: string
        content:
          type: string
    NoteResponse:
      allOf:
        - $ref: '#/components/schemas/Note'
        - type: object
          required:
            - id
          properties:
            id:
              type: string
              format: uuid
    NoteCreated:
      type: object
      required:
        - uri
      properties:
        uri:
          type: string
          format: url
  requestBodies:
    NoteRequest:
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Note'